Added methods to MyCircle

+++ src/sample/MyShape.java
@@ -1,34 +1,24 @@
 package sample;
 
 public class MyShape {
-    protected int x;
-    protected int y;
+    protected double x;
+    protected double y;
 
     public MyShape() {
         this.x = 0;
         this.y = 0;
     }
 
-    public MyShape(int x, int y) {
+    public MyShape(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
-    public int getX() {
-        return x;
-    }
+    public double getX() {  return x; }
 
-    public void setX(int x) {
-        this.x = x;
-    }
+    public void setX(double x) { this.x = x; }
 
-    public int getY() {
-        return y;
-    }
+    public double getY() { return y; }
 
-    public void setY(int y) {
-        this.y = y;
-    }
-
-
+    public void setY(double y) { this.y = y; }
 }
